동민5까지

+++ src/main/java/warmingUp/antifragile/post/controller/PostController.java
@@ -26,9 +26,12 @@ public class PostController {
 
     @PostMapping("/reviews")
     public ReturnOneDto<PostReadDto> createPost(@RequestBody PostCreateDto postCreateDto, HttpServletRequest request){
-        HttpSession session = request.getSession();
-        Long memberId = (Long)session.getAttribute("memberID");
         //로그인이 안되어 있을 때
+        HttpSession session = request.getSession(false);
+        if(session == null)
+            return new ReturnOneDto<>(null, "로그인을 하세요");
+
+        Long memberId = (Long)session.getAttribute("memberID");
         if(memberId == null)
             return new ReturnOneDto<>(null, "로그인을 하세요");
         //맴버 조회실패, 차가 없는 경우, 성공했을 경우
@@ -37,12 +40,14 @@ public class PostController {
 
     @DeleteMapping("/reviews/{postId}")
     public ReturnOneDto<PostReadDto> deletePost(@PathVariable Long postId, HttpServletRequest request){
-        HttpSession session = request.getSession();
-        Long memberId = (Long)session.getAttribute("memberID");
+        //로그인이 안되어 있을 때
+        HttpSession session = request.getSession(false);
+        if(session == null)
+            return new ReturnOneDto<>(null, "로그인을 하세요");
 
-        //로그인 안되어 있는 경우
+        Long memberId = (Long)session.getAttribute("memberID");
         if(memberId == null)
-            return new ReturnOneDto<>(null,"로그인을 하세요");
+            return new ReturnOneDto<>(null, "로그인을 하세요");
         // 그 밖의 경우
         return postService.deletePost(memberId,postId);
 
@@ -50,12 +55,14 @@ public class PostController {
 
     @PatchMapping("/reviews/{postId}")
     public ReturnOneDto<PostReadDto> modifyPost(@PathVariable Long postId,@RequestBody PostCreateDto postCreateDto, HttpServletRequest request){
-        HttpSession session = request.getSession();
-        Long memberId = (Long)session.getAttribute("memberID");
+        //로그인이 안되어 있을 때
+        HttpSession session = request.getSession(false);
+        if(session == null)
+            return new ReturnOneDto<>(null, "로그인을 하세요");
 
-        //로그인 안되어 있는 경우
+        Long memberId = (Long)session.getAttribute("memberID");
         if(memberId == null)
-            return new ReturnOneDto<>(null,"로그인을 하세요");
+            return new ReturnOneDto<>(null, "로그인을 하세요");
 
         return postService.modifyPost(memberId,postId, postCreateDto);
